requirements.txt
Flask
opencv-python
numpy
pytesseract
easyocr
torch
Pillow
transformers
google-generativeai